+++ NEWS.md
@@ -1,6 +1,7 @@
 ### crunch 1.18.1 (under development)
 * Changed the behavior of `changeCategoryID()` to only update numeric values of the category having its id changed when the id and the numeric value are the same.
 * The `autorollback` argument of `appendDataset()` has been deprecated. The option no longer has any effect and a warning will be printed to notify users about the deprecation.
+* `makeWeight` allows you to generate new weighting variables based on catagorical variables. Closes #80. 
 
 ## crunch 1.18.0